relocating code
