+++ internal/controller/monovertexrollout/monovertexrollout_controller.go
@@ -56,8 +56,6 @@ import (
 
 const (
 	ControllerMonoVertexRollout = "monovertex-rollout-controller"
-	TemplateMonoVertexName      = ".monovertex-name"
-	TemplateMonoVertexNamespace = ".monovertex-namespace"
 )
 
 // MonoVertexRolloutReconciler reconciles a MonoVertexRollout object
@@ -635,8 +633,8 @@ func (r *MonoVertexRolloutReconciler) makeMonoVertexDefinition(
 ) (*unstructured.Unstructured, error) {
 
 	args := map[string]interface{}{
-		TemplateMonoVertexName:      monoVertexName,
-		TemplateMonoVertexNamespace: monoVertexRollout.Namespace,
+		common.TemplateMonoVertexName:      monoVertexName,
+		common.TemplateMonoVertexNamespace: monoVertexRollout.Namespace,
 	}
 
 	monoVertexSpec, err := util.ResolveTemplatedSpec(monoVertexRollout.Spec.MonoVertex.Spec, args)
@@ -769,8 +767,8 @@ func (r *MonoVertexRolloutReconciler) GetDesiredRiders(rolloutObject ctlrcommon.
 			return desiredRiders, fmt.Errorf("rider definition could not converted to map: %w", err)
 		}
 		resolvedMap, err := util.ResolveTemplatedSpec(asMap, map[string]interface{}{
-			TemplateMonoVertexName:      monoVertexName,
-			TemplateMonoVertexNamespace: monoVertexRollout.Namespace,
+			common.TemplateMonoVertexName:      monoVertexName,
+			common.TemplateMonoVertexNamespace: monoVertexRollout.Namespace,
 		})
 		if err != nil {
 			return desiredRiders, err
